Crate tracing_android

Integration layer between tracing and Android logs.

This crate provides a MakeWriter and a Layer suitable for writing Android logs.

Usage

tracing_android::init(env!("CARGO_PKG_NAME"));

or in with custom options and combined with other layers

use tracing_subscriber::Registry;
use tracing_subscriber::fmt::FmtSpan;
use tracing_subscriber::prelude::*;

let android_layer = tracing_android::layer(env!("CARGO_PKG_NAME"))
    .with_span_events(FmtSpan::CLOSE)
    .with_thread_names(true);

let registry = Registry::default()
    .with(android_layer)
    .with(other_layer);

Cargo features

api-30: Enables support for Android API level 30 and source location information

json: Enables support for the JSON log format

log: Enables support for the JSON log format

Structs

AndroidLogMakeWriter

The writer produced by AndroidLogMakeWriter.

AndroidLogWriter

A MakeWriter suitable for writing Android logs.

Layer

A Layer that writes formatted representations of tracing events as Android logs.

Enums

Buffer

An Android log buffer.

Functions

init

Creates a Subscriber with the given tag and attempts to set it as the global default subscriber in the current scope, panicking if this fails.

layer

Returns a new formatting layer that can be composed with other layers to construct a Subscriber.

Type Definitions

Subscriber

A Subscriber that writes formatted representations of tracing events as Android logs.