Struct tracing_android::Layer

pub struct Layer<S, N = DefaultFields, E = Full>(_);

A Layer that writes formatted representations of tracing events as Android logs.

Implementations

impl<S> Layer<S> where
S: Subscriber + for<'a> LookupSpan<'a>,

pub fn new(tag: impl ToString) -> Self

Returns a new Layer with the given tag.

pub fn with_buffer(tag: impl ToString, buffer: Buffer) -> Self

Returns a new Layer with the given tag and using the given Android log buffer.

impl<N, E> Layer<Registry, N, E> where
Layer<Registry, N, Format<E, ()>, AndroidLogMakeWriter>: Layer<Registry>,

pub fn subscriber(self) -> Subscriber<N, E>

Returns a Subscriber by wrapping self in a Registry.

pub fn init(self) where
Subscriber<N, E>: Into<Dispatch>,

Converts self into a Subscriber and attempts to set it as the global default subscriber in the current scope, panicking if this fails.

Panics

This method panics if a global default subscriber has already been set, or if a log logger has already been set (when the “log” feature is enabled).

log compatibility

If the “log” feature flag is enabled, this will also attempt to initialize a log compatibility layer. This allows the subscriber to consume log::Records as though they were tracing Events.

impl<S, N, E> Layer<S, N, E> where
N: for<'writer> FormatFields<'writer> + 'static,

pub fn with_span_events(self, kind: FmtSpan) -> Self

See tracing_subscriber documentation

pub fn with_target(self, display_target: bool) -> Self

See tracing_subscriber documentation

pub fn with_thread_ids(self, display_thread_ids: bool) -> Self

See tracing_subscriber documentation

pub fn with_thread_names(self, display_thread_names: bool) -> Self

See tracing_subscriber documentation

pub fn compact(self) -> Layer<S, N, Compact>

See tracing_subscriber documentation

pub fn json(self) -> Layer<S, JsonFields, Json>

This is supported on crate feature json only.

See tracing_subscriber documentation

impl<S> Layer<S, JsonFields, Json>

pub fn flatten_event(self, flatten_event: bool) -> Self

This is supported on crate feature json only.

See tracing_subscriber documentation

pub fn with_current_span(self, display_current_span: bool) -> Self

This is supported on crate feature json only.

See tracing_subscriber documentation

pub fn with_span_list(self, display_span_list: bool) -> Self

This is supported on crate feature json only.

See tracing_subscriber documentation

impl<S, N, E> Layer<S, N, E>

pub fn fmt_fields<N2>(self, fmt_fields: N2) -> Layer<S, N2, E> where
N2: for<'writer> FormatFields<'writer> + 'static,

See tracing_subscriber documentation