Type Definition tracing_android::Subscriber

pub type Subscriber<N, E> = Layered<Layer<Registry, N, Format<E, ()>, AndroidLogMakeWriter>, Registry>;

A Subscriber that writes formatted representations of tracing events as Android logs.

This consists of a Layer wrapped in a Registry.